@@ -1,6 +1,6 @@
 {
   "name": "ng-pick-datetime",
-  "version": "7.0.1",
+  "version": "8.0.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
@@ -77,7 +77,7 @@
     "rxjs": "6.3.3",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "3.1.3",
+    "typescript": "^3.4.0",
     "web-animations-js": "^2.3.1",
     "zone.js": "0.8.26"
   }
