@@ -1782,8 +1782,8 @@ class StandardDateTimePicker {
     pickerMode = 'popup';
     opened = false;
     disabled = false;
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
 }
 
 @Component({
@@ -1798,8 +1798,8 @@ class RangeDateTimePicker {
     selectMode = 'range';
     pickerType = 'both';
     startAt = new Date(2020, JAN, 1);
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
 }
 
 @Component({
@@ -1818,7 +1818,7 @@ class MultiInputDateTimePicker {
     `,
 })
 class NoInputDateTimePicker {
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
 }
 
 @Component({
@@ -1830,7 +1830,7 @@ class NoInputDateTimePicker {
 class DateTimePickerWithStartAt {
     date = new Date(2020, JAN, 1);
     startDate = new Date(2010, JAN, 1);
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
 }
 
 @Component({
@@ -1844,7 +1844,7 @@ class DateTimePickerWithStartAt {
 class DateTimePickerWithStartView {
     date = new Date(2020, JAN, 1);
     startView = 'month';
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
 
     onMonthSelection() {
     }
@@ -1862,8 +1862,8 @@ class DateTimePickerWithStartView {
 class DateTimePickerWithNgModel {
     moment: Date[] | Date | null = null;
     selectMode = 'single';
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
 }
 
 @Component({
@@ -1876,9 +1876,9 @@ class DateTimePickerWithNgModel {
 })
 class DateTimePickerWithFormControl {
     formControl = new FormControl();
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
-    @ViewChild(OwlDateTimeTriggerDirective) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild(OwlDateTimeTriggerDirective, {static: true}) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
 }
 
 @Component({
@@ -1889,8 +1889,8 @@ class DateTimePickerWithFormControl {
     `,
 })
 class DateTimePickerWithTrigger {
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
 }
 
 @Component({
@@ -1902,9 +1902,9 @@ class DateTimePickerWithTrigger {
     `,
 })
 class DateTimePickerWithMinAndMaxValidation {
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
-    @ViewChild(OwlDateTimeTriggerDirective) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild(OwlDateTimeTriggerDirective, {static: true}) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
 
     date: Date | null;
     min: Date;
@@ -1921,9 +1921,9 @@ class DateTimePickerWithMinAndMaxValidation {
     `,
 })
 class DateTimePickerWithFilterValidation {
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
-    @ViewChild(OwlDateTimeTriggerDirective) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild(OwlDateTimeTriggerDirective, {static: true}) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
     date: Date;
     filter = ( date: Date ) => date.getDate() != 1;
 }
@@ -1940,9 +1940,9 @@ class DateTimePickerWithFilterValidation {
     `,
 })
 class DateTimePickerWithChangeAndInputEvents {
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
-    @ViewChild(OwlDateTimeTriggerDirective) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild(OwlDateTimeTriggerDirective, {static: true}) dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
 
     handleChange() {
     }
@@ -1968,8 +1968,8 @@ class DateTimePickerWithISOStrings {
     min = new Date(2017, JAN, 1).toISOString();
     max = new Date (2017, DEC, 31).toISOString();
     startAt = new Date(2017, JUL, 1).toISOString();
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
-    @ViewChild(OwlDateTimeInputDirective) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild(OwlDateTimeInputDirective, {static: true}) dateTimePickerInput: OwlDateTimeInputDirective<Date>;
 }
 
 @Component({
@@ -1982,5 +1982,5 @@ class DateTimePickerWithEvents {
     selected: Date | null = null;
     openedSpy = jasmine.createSpy('opened spy');
     closedSpy = jasmine.createSpy('closed spy');
-    @ViewChild('dt') dateTimePicker: OwlDateTimeComponent<Date>;
+    @ViewChild('dt', {static: true}) dateTimePicker: OwlDateTimeComponent<Date>;
 }
